api_server/requirements copy.txt
aiodns
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
annotated-types
anyio
asgiref
asttokens
attrs
backoff
bcrypt
beautifulsoup4
blinker
boto3
botocore
brotli
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
clip
colorama
coloredlogs
comm
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
deprecated
distro
dnspython
docstring_parser
duckduckgo-search
durationpy
einops
email_validator
executing
fastapi
fastapi-cli
filelock
flask
flatbuffers
fonttools
frozenlist
fsspec
ftfy
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
googlesearch-python
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
inquirerpy
ipykernel
ipython
ipython_pygments_lexers
itsdangerous
jedi
jinja2
jiter
jmespath
joblib
jsonpatch
jsonpointer
jupyter_client
jupyter_core
kiwisolver
kubernetes
langchain
langchain-aws
langchain-community
langchain-core
langchain-google-vertexai
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
markdown-it-py
markupsafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mmh3
monotonic
mpmath
msgpack
multidict
mypy-extensions
nest-asyncio
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
ollama
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
overrides
packaging
pandas
paramiko
parso
pdfminer.six
pexpect
pfzy
pillow
platformdirs
posthog
prettytable
primp
prompt_toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure_eval
py-cpuinfo
pyasn1
pyasn1_modules
pycares
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pygments
pymupdf
pynacl
pyparsing
pypika
pyproject_hooks
pysocks
python-dateutil
python-dotenv
python-multipart
pytz
pyyaml
pyzmq
PyMuPDF
psycopg2-binary
pytesseract
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rsa
runpod
s3transfer
safetensors
scikit-learn
scipy
selenium
sentence-transformers
shapely
shellingham
six
sniffio
sortedcontainers
soupsieve
sqlalchemy
stack-data
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
tomlkit
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu122
tornado
tqdm
tqdm-loggable
traitlets
transformers
trio
trio-websocket
triton
typer
typing-inspect
typing_extensions
tzdata
ujson
urllib3
uvicorn
uvloop
watchdog
watchfiles
wcwidth
webdriver-manager
websocket-client
websockets
werkzeug
wrapt
wsproto
yarl
zipp
zstandard